+++ lib/playbook/index.ts
@@ -1,0 +1,8 @@
+export * as Academic from "./academic";
+export * as Opportunities from "./opportunities";
+export * as Compass from "./compass";
+export * as Trust from "./trust";
+export * as Record from "./record";
+export * as Timeline from "./timeline";
+export * as Events from "./events";
+export * as Repositories from "./repositories";